requirements.txt
tensorflow>=2.13.0
keras>=2.13.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=9.5.0
scikit-learn>=1.3.0
jupyter>=1.0.0
ipython>=8.0.0
streamlit>=1.28.0
